Unify ulp metadata parser functions
Previously, the metadata parser routines were different between
toolside and libside. In this commit we unify these functions, as
there is no point in keeping them different now.
