+++ include/msg_queue.h
@@ -22,6 +22,8 @@
 #ifndef MSGQ_H
 #define MSGQ_H
 
+#include <stdarg.h>
+
 /** Define a 2Mb buffer for holding the messages.  */
 #define MSGQ_BUFFER_MAX (2 * 1024 * 1024)
 
@@ -77,11 +79,6 @@ struct msg_queue
 extern struct msg_queue __ulp_msg_queue;
 
 void msgq_push(const char *format, ...);
-
-/** Send a warning message to the message queue.  */
-#define MSGQ_WARN(fmt, ...) msgq_push("ulp: " fmt "\n", ##__VA_ARGS__)
-
-/** Send a debug message to the message queue, if debugging is enabled.  */
-#define MSGQ_DEBUG(fmt, ...) msgq_push("ulp: " fmt "\n", ##__VA_ARGS__)
+void msgq_vpush(const char *format, va_list);
 
 #endif /* MSGQ_H */
